+++ ng/nggui.cpp
@@ -175,32 +175,46 @@ const std::string* ng::Text::next_line () {
 
 ng::Tileset::Tileset () {
 	this->image = NULL;
+	this->grid.set(1.0, 1.0, 0.0, 0.0);
+	this->offset.set(0.0, 0.0);
 }
 
 ng::Tileset::~Tileset () {}
 
-void ng::Tileset::set (Image* image, const Space* space) {
+void ng::Tileset::set (Image* image) {
 	this->image = image;
-	this->space = *space;
-	this->offset.set2(0.0, 0.0);
+	this->grid.set(1.0, 1.0, image->w, image->h);
+	this->offset.set(0.0, 0.0);
 }
 
-void ng::Tileset::set (Image* image, const Rect* rect) {
+void ng::Tileset::set (Image* image, double c, double r) {
 	this->image = image;
-	this->space.set2(rect);
-	this->offset.set2(0.0, 0.0);
+	this->grid.set(ng::units(c, image->w), ng::units(r, image->h), c, r);
+	this->offset.set(0.0, 0.0);
 }
 
-void ng::Tileset::set_c (Image* image, const Rect* rect, double c, double r) {
+void ng::Tileset::set (Image* image, double x, double y, double c, double r) {
 	this->image = image;
-	this->space.set2_c(rect, c, r);
-	this->offset.set2(0.0, 0.0);
+	this->grid.set(x, y, ng::units(c, image->w), ng::units(r, image->h), c, r);
+	this->offset.set(0.0, 0.0);
 }
 
-void ng::Tileset::set_i (Image* image, const Rect* rect, double i, double j) {
+void ng::Tileset::set (Image* image, const Vec2* p, double c, double r) {
 	this->image = image;
-	this->space.set2_i(rect, i, j);
-	this->offset.set2(0.0, 0.0);
+	this->grid.set(p, ng::units(c, image->w), ng::units(r, image->h), c, r);
+	this->offset.set(0.0, 0.0);
+}
+
+void ng::Tileset::set (Image* image, const Grid2* grid) {
+	this->image = image;
+	this->grid = *grid;
+	this->offset.set(0.0, 0.0);
+}
+
+void ng::Tileset::absolute (Rect2* const rect) const {
+	rect->p.x += this->offset.x;
+	rect->p.y += this->offset.y;
+	this->grid.absolute(rect);
 }
 
 ng::Button::Button () {}
@@ -210,9 +224,13 @@ ng::Button::~Button () {}
 void ng::Button::set (const char* str, double text_scale, double x, double y, double w, double h) {
 	this->text.reset();
 	this->text.push_line(str);
-	this->rect.set2(x, y, w, h);
-	this->space.set2_i(&this->rect, h*text_scale*0.5, h*text_scale);
+	double i, j;
+	i = h*text_scale*0.5;
+	j = h*text_scale;
+	this->text_grid.set(x - (w*0.5), y - (h*0.5),
+		i, j, ng::columns(i, w), ng::columns(j, h));
 	this->text_color.set(255, 255, 255);
+	this->rect.set(x, y, w, h);
 	this->fill_color.set(0, 0, 0);
 	this->frame_color.set(255, 255, 255);
 }
@@ -221,15 +239,19 @@ void ng::Button::set (const char* str, double text_scale, const Color* text_colo
 double x, double y, double w, double h, const Color* fill, const Color* frame) {
 	this->text.reset();
 	this->text.push_line(str);
-	this->rect.set2(x, y, w, h);
-	this->space.set2_i(&this->rect, h*text_scale*0.5, h*text_scale);
+	double i, j;
+	i = h*text_scale*0.5;
+	j = h*text_scale;
+	this->text_grid.set(x - (w*0.5), y - (h*0.5),
+		i, j, ng::columns(i, w), ng::columns(j, h));
 	this->text_color = *text_color;
+	this->rect.set(x, y, w, h);
 	this->fill_color = *fill;
 	this->frame_color = *frame;
 }
 
-bool ng::Button::contains (const Vec* p) const {
-	return this->rect.contains2(p->x, p->y);
+bool ng::Button::contains (const Vec2* p) const {
+	return this->rect.contains(p);
 }
 
 ng::Label::Label () {}
@@ -239,218 +261,107 @@ ng::Label::~Label () {}
 void ng::Label::set (const char* str, double lines, double x, double y, double w, double h) {
 	this->text.reset();
 	this->text.push_line(str);
-	ng::Rect r;
-	r.set2(x, y, w, h);
-	this->space.set2_i(&r, (h*0.5)/lines, h/lines);
+	this->text_grid.set(x - (w*0.5), y - (h*0.5),
+		ng::units(lines*2.0, w), ng::units(lines, h), lines*2.0, lines);
 	this->text_color.set(255, 255, 255);
+	this->rect.set(x, y, w, h);
 }
 
 void ng::Label::set (const char* str, double lines, const Color* text_color,
 double x, double y, double w, double h) {
 	this->text.reset();
 	this->text.push_line(str);
-	ng::Rect r;
-	r.set2(x, y, w, h);
-	this->space.set2_i(&r, (h*0.5)/lines, h/lines);
+	this->text_grid.set(x - (w*0.5), y - (h*0.5),
+		ng::units(lines*2.0, w), ng::units(lines, h), lines*2.0, lines);
 	this->text_color = *text_color;
+	this->rect.set(x, y, w, h);
 }
 
-bool ng::Label::contains (const Vec* p) const {
-	return this->space.rect.contains2(p->x, p->y);
+bool ng::Label::contains (const Vec2* p) const {
+	return this->rect.contains(p);
 }
 
 ng::Canvas::Canvas () {
 	this->graphics = NULL;
 	this->parent = NULL;
 	this->root = false;
-	this->relative = false;
 }
 
 ng::Canvas::~Canvas () {}
 
-void ng::Canvas::set (Graphics* graphics) {
+void ng::Canvas::set (Graphics* graphics, double w, double h) {
 	this->graphics = graphics;
 	this->parent = NULL;
-	this->space.set2(&graphics->rect);
 	this->root = true;
-	this->relative = false;
+	this->space.set(w * 0.5, h * 0.5, 1.0, 1.0);
+	this->w = w;
+	this->h = h;
 }
 
-void ng::Canvas::set_c (Graphics* graphics, double c, double r) {
+void ng::Canvas::set (Graphics* graphics, double x, double y, double w, double h) {
 	this->graphics = graphics;
 	this->parent = NULL;
-	this->space.set2_c(&graphics->rect, c, r);
 	this->root = true;
-	this->relative = false;
+	this->space.set(x, y, 1.0, 1.0);
+	this->w = w;
+	this->h = h;
 }
 
-void ng::Canvas::set_i (Graphics* graphics, double i, double j) {
+void ng::Canvas::set (Graphics* graphics, const Space2* space, double w, double h) {
 	this->graphics = graphics;
 	this->parent = NULL;
-	this->space.set2_i(&graphics->rect, i, j);
 	this->root = true;
-	this->relative = false;
-}
-
-void ng::Canvas::set (Canvas* canvas, const Space* space) {
-	this->graphics = canvas->graphics;
-	this->parent = canvas;
 	this->space = *space;
-	this->root = false;
-	this->relative = true;
+	this->w = w;
+	this->h = h;
 }
 
 void ng::Canvas::set (Canvas* canvas, double x, double y, double w, double h) {
 	this->graphics = canvas->graphics;
 	this->parent = canvas;
-	this->space.set2(x, y, w, h);
 	this->root = false;
-	this->relative = true;
+	this->space.set(x, y, 1.0, 1.0);
+	this->w = w;
+	this->h = h;
 }
 
-void ng::Canvas::set_c (Canvas* canvas, double x, double y, double w, double h, double c, double r) {
+void ng::Canvas::set (Canvas* canvas, const Space2* space, double w, double h) {
 	this->graphics = canvas->graphics;
 	this->parent = canvas;
-	this->space.set2_c(x, y, w, h, c, r);
 	this->root = false;
-	this->relative = true;
+	this->space = *space;
+	this->w = w;
+	this->h = h;
 }
 
-void ng::Canvas::set_i (Canvas* canvas, double x, double y, double w, double h, double i, double j) {
-	this->graphics = canvas->graphics;
-	this->parent = canvas;
-	this->space.set2_i(x, y, w, h, i, j);
-	this->root = false;
-	this->relative = true;
-}
-
-// scale from parent/graphics to this.
-void ng::Canvas::scale_in (Vec* const vec) {
-	if (this->relative) {
-		vec->absolute_to_relative2(&this->space);
-	} else {
-		Vec s;
-		if (this->root) {
-			s.set2(&this->graphics->rect, &this->space);
-		} else {
-			s.set2(&this->parent->space, &this->space);
-		}
-		vec->scale2(s.x, s.y);
-	}
-}
-
-void ng::Canvas::scale_in (Rect* const rect) {
-	if (this->relative) {
-		rect->absolute_to_relative2(&this->space);
-	} else {
-		Vec s;
-		if (this->root) {
-			s.set2(&this->graphics->rect, &this->space);
-		} else {
-			s.set2(&this->parent->space, &this->space);
-		}
-		rect->scale2(s.x, s.y);
-	}
-}
-
-void ng::Canvas::scale_in (Space* const space) {
-	if (this->relative) {
-		space->absolute_to_relative2(&this->space);
-	} else {
-		Vec s;
-		if (this->root) {
-			s.set2(&this->graphics->rect, &this->space);
-		} else {
-			s.set2(&this->parent->space, &this->space);
-		}
-		space->scale2(s.x, s.y);
-	}
-}
-
-// scale from this to parent/graphics.
-void ng::Canvas::scale_out (Vec* const vec) {
-	if (this->relative) {
-		vec->relative_to_absolute2(&this->space);
-	} else {
-		Vec s;
-		if (this->root) {
-			s.set2(&this->space, &this->graphics->rect);
-		} else {
-			s.set2(&this->space, &this->parent->space);
-		}
-		vec->scale2(s.x, s.y);
-	}
-}
-
-void ng::Canvas::scale_out (Rect* const rect) {
-	if (this->relative) {
-		rect->relative_to_absolute2(&this->space);
-	} else {
-		Vec s;
-		if (this->root) {
-			s.set2(&this->space, &this->graphics->rect);
-		} else {
-			s.set2(&this->space, &this->parent->space);
-		}
-		rect->scale2(s.x, s.y);
-	}
-}
-
-void ng::Canvas::scale_out (Space* const space) {
-	if (this->relative) {
-		space->relative_to_absolute2(&this->space);
-	} else {
-		Vec s;
-		if (this->root) {
-			s.set2(&this->space, &this->graphics->rect);
-		} else {
-			s.set2(&this->space, &this->parent->space);
-		}
-		space->scale2(s.x, s.y);
-	}
+// Get the bounding box for this canvas, relative to this space.
+void ng::Canvas::get_rect (Rect2* const rect) const {
+	rect->set(0.0, 0.0, this->w, this->h);
 }
 
 // Given event mouse point on root canvas, find mouse point on this canvas.
-void ng::Canvas::get_mouse (Vec* const mouse) {
+void ng::Canvas::get_mouse (Vec2* const mouse) const {
 	if (!this->root) {
 		this->parent->get_mouse(mouse);
 	}
-	this->scale_in(mouse);
-}
-
-// Draw this->space on parent.
-void ng::Canvas::clear () {
-	if (this->relative && this->root) {
-		this->graphics->draw_rect(&this->space.rect, ng::DrawFill);
-	} else if (this->relative && !this->root) {
-		this->parent->draw_rect(&this->space.rect, ng::DrawFill);
-	} else if (!this->relative && this->root) {
-		this->graphics->clear();
-	} else {
-		this->parent->clear();
-	}
-}
-
-// Usually does nothing.
-// Drawing root canvas calls graphics draw().
-void ng::Canvas::draw () {
-	if (this->root) {
-		this->graphics->draw();
-	}
+	this->space.relative(mouse);
 }
 
 // Graphics primitives
 void ng::Canvas::draw_image (Image* const image) {
-	this->draw_image(image, &image->rect, &this->space.rect);
+	Rect2 src(image->w * 0.5, image->h * 0.5, image->w, image->h);
+	Rect2 dest(0.0, 0.0, this->w, this->h);
+	this->draw_image(image, &src, &dest);
 }
 
-void ng::Canvas::draw_image (Image* const image, const Rect* dest) {
-	this->draw_image(image, &image->rect, dest);
+void ng::Canvas::draw_image (Image* const image, const Rect2* dest) {
+	Rect2 src(image->w * 0.5, image->h * 0.5, image->w, image->h);
+	this->draw_image(image, &src, dest);
 }
 
-void ng::Canvas::draw_image (Image* const image, const Rect* src, const Rect* dest) {
-	Rect d = *dest;
-	this->scale_out(&d);
+void ng::Canvas::draw_image (Image* const image, const Rect2* src, const Rect2* dest) {
+	Rect2 d = *dest;
+	this->space.absolute(&d);
 	
 	if (this->root) {
 		this->graphics->draw_image(image, src, &d);
@@ -459,9 +370,9 @@ void ng::Canvas::draw_image (Image* const image, const Rect* src, const Rect* de
 	}
 }
 
-void ng::Canvas::draw_rect (const Rect* dest, int draw) {
-	Rect d = *dest;
-	this->scale_out(&d);
+void ng::Canvas::draw_rect (const Rect2* dest, int draw) {
+	Rect2 d = *dest;
+	this->space.absolute(&d);
 	
 	if (this->root) {
 		this->graphics->draw_rect(&d, draw);
@@ -470,94 +381,74 @@ void ng::Canvas::draw_rect (const Rect* dest, int draw) {
 	}
 }
 
-void ng::Canvas::draw_line (double x1, double y1, double x2, double y2) {
-	Vec v1, v2;
-	v1.set2(x1, y1);
-	v2.set2(x2, y2);
-	this->scale_out(&v1);
-	this->scale_out(&v2);
+void ng::Canvas::draw_line (const Line2* line) {
+	Line2 l = *line;
+	this->space.absolute(&l);
 	
 	if (this->root) {
-		this->graphics->draw_line(v1.x, v1.y, v2.x, v2.y);
+		this->graphics->draw_line(&l);
 	} else {
-		this->parent->draw_line(v1.x, v1.y, v2.x, v2.y);
+		this->parent->draw_line(&l);
 	}
 }
 
-void ng::Canvas::draw_point (double x, double y) {
-	Vec v;
-	v.set2(x, y);
-	this->scale_out(&v);
+void ng::Canvas::draw_point (const Vec2* p) {
+	Vec2 v = *p;
+	this->space.absolute(&v);
 	
 	if (this->root) {
-		this->graphics->draw_point(v.x, v.y);
+		this->graphics->draw_point(&v);
 	} else {
-		this->parent->draw_point(v.x, v.y);
+		this->parent->draw_point(&v);
 	}
 }
 
 // Advanced graphics
-void ng::Canvas::draw_text (Tileset* const tileset, Text* const text, const Space* dest) {
-	if (this->root) {
-		Rect s, tile, d;
-		s.set2(0.0, 0.0, 1.0, 1.0);
-		tile.set2(0.0, 0.0, 1.0, 1.0);
-		
-		int text_line, text_i;
-		text_line = text->line;
-		text_i = text->i;
-		char ch;
-		
-		double c;
-		while ((ch = text->next()) != '\0') {
-			if (ch == '\n') {
-				tile.x = 0.0;
-				tile.y += 1.0;
-				if (tile.y >= dest->r) {
-					return;
-				}
-				continue;
+void ng::Canvas::draw_text (Tileset* const tileset, Text* const text, const Grid2* dest) {
+	Rect2 s, tile, d;
+	s.set(0.5, 0.5, 1.0, 1.0);
+	tile.set(0.5, 0.5, 1.0, 1.0);
+	
+	int text_line, text_i;
+	text_line = text->line;
+	text_i = text->i;
+	char ch;
+	
+	double c;
+	while ((ch = text->next()) != '\0') {
+		if (ch == '\n') {
+			tile.p.x = 0.5;
+			tile.p.y += 1.0;
+			if (tile.p.y >= dest->r) {
+				return;
 			}
-			
-			c = static_cast<double>(ch);
-			s.x = fmod(c, tileset->space.c);
-			s.y = floor(c / tileset->space.c);
-			d = tile;
-			d.relative_to_absolute2(dest);
-			this->draw_tile(tileset, &s, &d);
-			
-			tile.x += 1.0;
-			if (tile.x >= dest->c) {
-				tile.x = 0.0;
-				tile.y += 1.0;
-				if (tile.y >= dest->r) {
-					return;
-				}
-			}
+			continue;
 		}
 		
-		text->set_cursor(text_line, text_i);
+		c = static_cast<double>(ch);
+		s.p.x = fmod(c, tileset->grid.c) + 0.5;
+		s.p.y = floor(c / tileset->grid.c) + 0.5;
+		d = tile;
+		dest->absolute(&d);
+		this->draw_tile(tileset, &s, &d);
 		
-	} else {
-		Space d = *dest;
-		this->scale_out(&d);
-		this->parent->draw_text(tileset, text, &d);
+		tile.p.x += 1.0;
+		if (tile.p.x >= dest->c) {
+			tile.p.x = 0.5;
+			tile.p.y += 1.0;
+			if (tile.p.y >= dest->r) {
+				return;
+			}
+		}
 	}
+	
+	text->set_cursor(text_line, text_i);	
 }
 
-void ng::Canvas::draw_tile (Tileset* const tileset, const Rect* src, const Rect* dest) {
-	Rect d = *dest;
-	this->scale_out(&d);
-	
-	if (this->root) {
-		Rect s = *src;
-		s.x += tileset->offset.x;
-		s.y += tileset->offset.y;
-		s.relative_to_absolute2(&tileset->space);
-		this->graphics->draw_image(tileset->image, &s, &d);
-	} else {
-		this->parent->draw_tile(tileset, src, &d);
-	}
+void ng::Canvas::draw_tile (Tileset* const tileset, const Rect2* src, const Rect2* dest) {
+	Rect2 s = *src;
+	tileset->absolute(&s);
+	this->draw_image(tileset->image, &s, dest);
 }
 
 // Gui elements
@@ -567,12 +458,12 @@ void ng::Canvas::draw_button (Tileset* const tileset, Button* const button) {
 	this->graphics->set_color(&button->frame_color);
 	this->draw_rect(&button->rect, ng::DrawFrame);
 	tileset->image->set_color(&button->text_color);
-	this->draw_text(tileset, &button->text, &button->space);
+	this->draw_text(tileset, &button->text, &button->text_grid);
 }
 
 void ng::Canvas::draw_label (Tileset* const tileset, Label* const label) {
 	tileset->image->set_color(&label->text_color);
-	this->draw_text(tileset, &label->text, &label->space);
+	this->draw_text(tileset, &label->text, &label->text_grid);
 }